# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI小说平台 - 拆书系统 API
  version: 1.0.0
  summary: 面向小B端的 AI 小说拆书系统接口契约
  description: |
    提供小说上传、处理状态跟踪、纲树查询、纲详情查看、错误日志查询与实时进度推送能力。
    本契约基于现有《API接口设计文档》整理为 OpenAPI 文档，用于前后端联调、Mock Server 生成与后续 CI 契约校验。
  contact:
    name: AI小说平台开发团队
    url: https://github.com/example/ai-novel-platform
servers:
  - url: /api/v1
    description: 相对路径（同域部署）
...
paths:
...
  /outlines/{outlineId}:
    get:
      tags: [Outlines]
      summary: 获取纲详情
      description: 根据纲ID查询完整的纲内容，包含章纲（含风格、节奏、场景模板）或通用纲结构化数据。
      operationId: getOutlineDetail
      parameters:
        - $ref: '#/components/parameters/OutlineId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutlineDetailResponse'
        '404':
          description: 纲不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /outlines/{outlineId}/copy:
    post:
      tags: [Outlines]
      summary: 获取可复制的纲内容
      description: 由服务端组装统一文案，便于前端一键复制。
      operationId: copyOutlineContent
      parameters:
        - $ref: '#/components/parameters/OutlineId'
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [text, json, markdown]
            default: text
          description: 导出格式，当前实现支持 text/json/markdown 三种格式
      responses:
        '200':
          description: 生成复制内容成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CopyOutlineResponse'
        '404':
          description: 纲不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: 获取任务状态
      description: 查询拆书任务的实时处理状态，包含当前阶段、各阶段进度百分比、已完成章节数及预估剩余时间。
      operationId: getTaskStatus
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatusResponse'
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tasks/{taskId}/errors:
    get:
      tags: [Tasks]
      summary: 获取任务错误日志
      description: 分页查询指定任务在处理过程中产生的错误记录，包含错误类型、所属阶段、发生时间等详细信息。
      operationId: getTaskErrors
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskErrorsResponse'
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
...
    OutlineId:
      name: outlineId
      in: path
      required: true
      schema:
        type: string
      description: 纲ID
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
      description: 任务ID
  schemas:
    ApiEnvelope:
      type: object
      required: [success, timestamp]
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
    ErrorInfo:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - INVALID_FILE_FORMAT
            - FILE_TOO_LARGE
            - ENCODING_DETECTION_FAILED
            - BOOK_NOT_FOUND
            - OUTLINE_NOT_FOUND
            - TASK_NOT_FOUND
            - PROCESSING_FAILED
            - LLM_ERROR
            - INTERNAL_ERROR
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          required: [success, data, error, timestamp]
          properties:
            success:
              type: boolean
              const: false
            data:
              type: object
              nullable: true
              additionalProperties: true
            error:
              $ref: '#/components/schemas/ErrorInfo'
...
    ProcessingStage:
      type: string
      enum:
        - FILE_UPLOAD
        - TEXT_PREPROCESS
        - CHAPTER_OUTLINE
        - COARSE_OUTLINE
        - MAIN_OUTLINE
        - WORLD_OUTLINE
    OutlineType:
      type: string
      enum: [CHAPTER, COARSE, MAIN, WORLD]
    OutlineStatus:
      type: string
      enum: [PENDING, PROCESSING, COMPLETED, FAILED]
...
    ChapterOutlineContent:
      type: object
      properties:
        generalStyle:
          type: string
        globalVisualRhythm:
          type: string
        settingsTemplate:
          type: object
          additionalProperties: true
        plotStyleIntegration:
          type: object
          additionalProperties: true
        summary:
          type: string
    GenericOutlineContent:
      type: object
      additionalProperties: true
    OutlineDetailData:
      type: object
      required: [outlineId, bookId, outlineType, status, content, summary, createdAt]
      properties:
        outlineId:
          type: string
        bookId:
          type: string
        outlineType:
          $ref: '#/components/schemas/OutlineType'
        chapterIndex:
          type: integer
        status:
          $ref: '#/components/schemas/OutlineStatus'
        content:
          oneOf:
            - $ref: '#/components/schemas/ChapterOutlineContent'
            - $ref: '#/components/schemas/GenericOutlineContent'
        summary:
          type: string
        createdAt:
          type: string
          format: date-time
    OutlineDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          required: [success, data, timestamp]
          properties:
            success:
              type: boolean
              const: true
            data:
              $ref: '#/components/schemas/OutlineDetailData'
    CopyOutlineData:
      type: object
      required: [outlineId, outlineType, copyContent, copyFormat]
      properties:
        outlineId:
          type: string
        outlineType:
          $ref: '#/components/schemas/OutlineType'
        copyContent:
          type: string
        copyFormat:
          type: string
          enum: [text, json, markdown]
    CopyOutlineResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          required: [success, data, timestamp]
          properties:
            success:
              type: boolean
              const: true
            data:
              $ref: '#/components/schemas/CopyOutlineData'
    StageProgress:
      type: object
      required:
        [FILE_UPLOAD, TEXT_PREPROCESS, CHAPTER_OUTLINE, COARSE_OUTLINE, MAIN_OUTLINE, WORLD_OUTLINE]
      properties:
        FILE_UPLOAD:
          type: integer
        TEXT_PREPROCESS:
          type: integer
        CHAPTER_OUTLINE:
          type: integer
        COARSE_OUTLINE:
          type: integer
        MAIN_OUTLINE:
          type: integer
        WORLD_OUTLINE:
          type: integer
    TaskStatusData:
      type: object
      required:
        [taskId, bookId, status, currentStage, stageProgress, totalChapters, completedChapters, errorCount, startTime]
      properties:
        taskId:
          type: string
        bookId:
          type: string
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, FAILED]
        currentStage:
          $ref: '#/components/schemas/ProcessingStage'
        stageProgress:
          $ref: '#/components/schemas/StageProgress'
        totalChapters:
          type: integer
        completedChapters:
          type: integer
        errorCount:
          type: integer
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          nullable: true
          format: date-time
        estimatedTimeRemaining:
          type: integer
          nullable: true
          description: 预计剩余秒数
    TaskStatusResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          required: [success, data, timestamp]
          properties:
            success:
              type: boolean
              const: true
            data:
              $ref: '#/components/schemas/TaskStatusData'
    TaskErrorItem:
      type: object
      required: [errorId, stage, errorType, errorMessage, timestamp]
      properties:
        errorId:
          type: string
        stage:
          $ref: '#/components/schemas/ProcessingStage'
        chapterIndex:
          type: integer
          nullable: true
        errorType:
          type: string
        errorMessage:
          type: string
        timestamp:
          type: string
          format: date-time
    TaskErrorsData:
      type: object
      required: [taskId, totalErrors, errors]
      properties:
        taskId:
          type: string
        totalErrors:
          type: integer
        errors:
          type: array
          items:
            $ref: '#/components/schemas/TaskErrorItem'
    TaskErrorsResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          required: [success, data, timestamp]
          properties:
            success:
              type: boolean
              const: true
            data:
              $ref: '#/components/schemas/TaskErrorsData'
...